# Conflicts:
#	code/go/0chain.net/smartcontract/dbs/event/miner.go
#	code/go/0chain.net/smartcontract/minersc/benchmark_setup.go

+++ code/go/0chain.net/smartcontract/zcnsc/handler.go
@@ -153,7 +153,7 @@ func toAuthorizerResponse(auth *event.Authorizer) *authorizerResponse {
 		Fee:             auth.Fee,
 		Latitude:        auth.Latitude,
 		Longitude:       auth.Longitude,
-		LastHealthCheck: auth.LastHealthCheck,
+		LastHealthCheck: int64(auth.LastHealthCheck),
 		DelegateWallet:  auth.DelegateWallet,
 		MinStake:        auth.MinStake,
 		MaxStake:        auth.MaxStake,
